adlib bank op2  midi adlib drv  refactor (#191)
* refactor OP2File, extract audio::opl::banks:Op2Bank_t Op2BankInstrument_t

* replace op2file with audio op2bank

* [WIP] Op2Bank_t -> OP2Bank

* OP2Bank class

* [WIP] refactor

* [wip]

* enable opl callback

* code rev

* utils opl test suite

* fix linux build

* disable test utils opl detectOPL as not working in CI linux & Mac

* std:share_ptr code rev const reference

* code rev

* code rev warnings

+++ sdl2-sonic-drivers/src/files/VOCFile.hpp
@@ -18,8 +18,8 @@ namespace files
 
         const std::string getVersion() const noexcept;
         const int getChannels() const noexcept;
-        const int getSampleRate() const noexcept;
-        const int getBitsDepth() const noexcept;
+        const uint32_t getSampleRate() const noexcept;
+        const uint8_t getBitsDepth() const noexcept;
         const int getDataSize() const noexcept;
         const std::shared_ptr<uint8_t[]> getData() const noexcept;
